*invalid relationship remove fix

+++ src/webdoc/gui/WNeuerPartner.java
@@ -193,8 +193,10 @@ public class WNeuerPartner extends JInternalFrame {
 			public void actionPerformed(ActionEvent e) {
 				logger.debug("recived delete command");
 				try {
-					Database.removeRelationship(id, JListTiere.getSelectedValue().getID());
-					loadAnimals();
+					if(JListTiere.getSelectedValue().getID() > 0){
+						Database.removeRelationship(id, JListTiere.getSelectedValue().getID());
+						loadAnimals();
+					}
 				} catch (SQLException e1) {
 					GUIManager.showDBErrorDialog(getParent(), Database.DBExceptionConverter(e1, true));
 				}
